refactor: update auth system and related components - Replace User with AuthUser in types - Update auth service implementations - Update auth routes and components - Remove deprecated auth.service.ts - Update signin and signup forms - Update API client - Update dashboard and signin pages
