+++ src/app/api/auth/signin/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { AuthService } from '@/lib/services/auth.service'
+import { AuthService } from '@/lib/services/supaservices'
 import { SigninRequest, ApiResponse } from '@/lib/types'
 
 export async function POST(request: NextRequest) {
