+++ src/components/signup/signup-form.tsx
@@ -7,7 +7,7 @@ import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/com
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { UserRole } from "@/lib/supabase"
-import { SupaService } from "@/lib/services/auth.service"
+import { SupaService } from "@/lib/services/supaservices"
 import { ArrowLeft, Loader2 } from "lucide-react"
 
 interface SignupFormProps {
